requirements.txt
langchain==0.1.0
langgraph==0.0.20
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.111.0
pydantic==2.5.2
email-validator==2.1.0
python-dotenv==1.0.0
streamlit==1.28.0
